refactor : AppTest와 acceptionTest 통합
- 더티 컨텍스트는 일단 주석을 통해 비활성화 해놓음.
