--- back/babble/src/test/java/gg/babble/babble/AcceptanceTest.java
@@ -1,22 +1,0 @@
-package gg.babble.babble;
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ActiveProfiles;
-
-@ActiveProfiles("test")
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class AcceptanceTest {
-
-    @LocalServerPort
-    protected int port;
-
-    @BeforeEach
-    public void setUp() {
-        RestAssured.port = port;
-    }
-}
