+++ back/babble/src/test/java/gg/babble/babble/ApplicationTest.java
@@ -1,12 +1,23 @@
 package gg.babble.babble;
 
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
 @ActiveProfiles("test")
 @Transactional
-@SpringBootTest
+//@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class ApplicationTest {
 
+    @LocalServerPort
+    protected int port;
+
+    @BeforeEach
+    public void setUp() {
+        RestAssured.port = port;
+    }
 }
