+++ back/babble/src/test/java/gg/babble/babble/controller/WebSocketTest.java
@@ -2,7 +2,7 @@ package gg.babble.babble.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gg.babble.babble.AcceptanceTest;
+import gg.babble.babble.ApplicationTest;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-public class WebSocketTest extends AcceptanceTest {
+public class WebSocketTest extends ApplicationTest {
 
     @Override
     @BeforeEach
