+++ back/babble/src/test/java/gg/babble/babble/restdocs/TagApiDocumentTest.java
@@ -12,7 +12,7 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.response
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import gg.babble.babble.AcceptanceTest;
+import gg.babble.babble.ApplicationTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -26,7 +26,7 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
 @ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
-public class TagApiDocumentTest extends AcceptanceTest {
+public class TagApiDocumentTest extends ApplicationTest {
 
     private MockMvc mockMvc;
